+++ Modules/NetworkingDsc/Examples/Resources/NetIPInterface/1-ConfigureNetIPInterface.ps1
@@ -1,0 +1,30 @@
+<#
+    .EXAMPLE
+    This example enables the following settings on the IPv4 network interface with alias
+    'Ethernet':
+    - AdvertiseDefaultRoute
+    - Avertising
+    - AutomaticMetric
+    - ClampMss
+    - Forwarding
+    - IgnoreDefaultRoute
+#>
+Configuration Example
+{
+    Import-DscResource -Module NetworkingDsc
+
+    Node localhost
+    {
+        NetIPInterface EnableSettings
+        {
+            InterfaceAlias        = 'Ethernet'
+            AddressFamily         = 'IPv4'
+            AdvertiseDefaultRoute = 'Enabled'
+            Advertising           = 'Enabled'
+            AutomaticMetric       = 'Enabled'
+            ClampMss              = 'Enabled'
+            Forwarding            = 'Enabled'
+            IgnoreDefaultRoutes   = 'Enabled'
+        }
+    }
+}