+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Added .VSCode settings for applying DSC PSSA rules - fixes [Issue #357](https://github.com/PowerShell/NetworkingDsc/issues/357).
 - Updated LICENSE file to match the Microsoft Open Source Team standard - fixes
   [Issue #363](https://github.com/PowerShell/NetworkingDsc/issues/363)
+- MSFT_NetIPInterface:
+  - Added a new resource for configuring the IP interface settings for a network
+    interface.
 
 ## 6.1.0.0